♻️ Split Heroes slices into per-use-case folders
Each Heroes use case now lives in its own folder with one file per
class (Request, Response, Endpoint, Validator, Summary). Namespaces
follow the folder, e.g. Features.Heroes.CreateHero.
