+++ src/WebApi/Features/Heroes/CreateHero/CreateHeroRequest.cs
@@ -1,0 +1,9 @@
+namespace SSW.VerticalSliceArchitecture.Features.Heroes.CreateHero;
+
+public record CreateHeroRequest(
+    string Name,
+    string Alias,
+    IEnumerable<CreateHeroRequest.HeroPowerDto> Powers)
+{
+    public record HeroPowerDto(string Name, int PowerLevel);
+}
