+++ src/WebApi/Features/Heroes/CreateHero/CreateHeroResponse.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Heroes.CreateHero;
+
+public record CreateHeroResponse(Guid Id);
